+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ server.error.include-message=always
 spring.datasource.username=root
 spring.datasource.password=********
 spring.datasource.url=jdbc:mysql://localhost:3306/fixturedb?serverTimezone=UTC
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=validate
 
 #Keep the connection alive if idle for long time (needed for production)
 spring.datasource.tomcat.test-while-idle=true
