+++ src/main/resources/templates/mainFixturesView.html
@@ -84,6 +84,7 @@
                             <form th:action="@{setOK-fixtureButton}" th:object="${newFixture}" method="post">
                                 <input name="id" type="hidden" th:value="${fixture.getId()}">
                                 <button name="action" type="submit" class="btn btn-primary btn-block">SetOK</button>
+                               <label> id="setOKdescr" ></label>
                                 <input style='width:100%' id="setOKdescr" type="text" th:field="*{descriptionOfChange}">
                             </form>
                         </div>
@@ -94,7 +95,7 @@
                             <input name="id" type="hidden" th:value="${fixture.getId()}">
 
                             <button name="action" type="submit" class="btn btn-primary btn-block">SetNOK</button>
-                            <input style='width:100%' id="setNOKdescr" type="text" th:field="*{descriptionOfChange}">
+                            <input style='width:100%' id="setNokdescr" type="text" th:field="*{descriptionOfChange}">
 
                         </form>
                     </td>
